requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Machine learning & clustering
scikit-learn>=1.3.0

# Keyphrase extraction (deterministic NLP)
yake>=0.4.8
rake-nltk>=1.0.6
nltk>=3.8.0

# Embeddings (multilingual, non-generative)
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Dimensionality reduction & visualization
umap-learn>=0.5.0
plotly>=5.15.0

# YAML parsing for playbooks
pyyaml>=6.0

# API server for remote access
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database (Neon PostgreSQL)
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
bcrypt>=4.1.0

# Kafka event streaming (optional — set KAFKA_BOOTSTRAP_SERVERS to enable)
aiokafka>=0.10.0